rename: 로딩페이지 파일위치 페이지스폴더로 이동 (#16)
