+++ src/pages/LoadingPage.jsx
